mudando as primeiras permissoes

+++ meu_projeto/carros/views.py
@@ -3,31 +3,47 @@ from django.contrib import messages
 from .forms import CarroForm
 from .models import Carro
 from django.views.decorators.http import require_POST
+from django.contrib.auth.decorators import login_required
 
+@login_required
 def carro_criar(request):
+    if request.user.is_admin is False:
+        messages.error(request, "Apenas administradores podem criar carros.")
+        return redirect("carros:disponiveis")
+    
     if request.method == "POST":
         form = CarroForm(request.POST, request.FILES)
         if form.is_valid():
-            carro = form.save()  # ⬅️ grava no banco
+            carro = form.save()
             messages.success(request, f"Carro {carro.marca} {carro.modelo} cadastrado!")
-            return redirect("carros:disponiveis")  # ⬅️ vai pra listagem
+            return redirect("carros:disponiveis")
     else:
         form = CarroForm()
     return render(request, "carros/cadastrar.html", {"form": form})
 
+@login_required
 def carros_disponiveis(request):
     carros = Carro.objects.filter(locatario__isnull=True).order_by("marca", "modelo")
     return render(request, "carros/disponiveis.html", {"carros": carros})
 
-
+@login_required
 @require_POST
 def carro_excluir(request, placa):   # use (request, pk) se for por id
+    if request.user.is_admin is False:
+        messages.error(request, "Apenas administradores podem excluir carros.")
+        return redirect("carros:disponiveis")
+    
     carro = get_object_or_404(Carro, placa=placa)  # ou get_object_or_404(Carro, pk=pk)
     carro.delete()
     messages.success(request, "Carro excluído com sucesso.")
     return redirect("carros:disponiveis")
 
+@login_required
 def carro_editar(request, placa):
+    if request.user.is_admin is False:
+        messages.error(request, "Apenas administradores podem editar carros.")
+        return redirect("carros:disponiveis")
+    
     carro = get_object_or_404(Carro, placa=placa)
     if request.method == "POST":
         form = CarroForm(request.POST, request.FILES, instance=carro)
